Création des traits Idtrait.php et DateTraits.php.
