+++ config/packages/ramsey_uuid_doctrine.yaml
@@ -1,0 +1,4 @@
+doctrine:
+    dbal:
+        types:
+            uuid: 'Ramsey\Uuid\Doctrine\UuidType'
