+++ src/Entity/Traits/DateTrait.php
@@ -1,0 +1,49 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: training
+ * Date: 1/30/20
+ * Time: 2:13 PM
+ */
+
+namespace App\Entity\Traits;
+
+
+trait DateTrait
+{
+    /**
+     * @ORM\Column(type="datetime_immutable")
+     */
+    private $createdDate;
+
+    /**
+     * @ORM\Column(type="datetime_immutable", nullable=true)
+     */
+    private $updatedDate;
+
+    public function getCreatedDate(): ?\DateTimeImmutable
+    {
+        return $this->createdDate;
+    }
+
+    public function setCreatedDate(\DateTimeImmutable $createdDate): self
+    {
+        $this->createdDate = $createdDate;
+
+        return $this;
+    }
+
+    public function getUpdatedDate(): ?\DateTimeImmutable
+    {
+        return $this->updatedDate;
+    }
+
+    public function setUpdatedDate(?\DateTimeImmutable $updatedDate): self
+    {
+        $this->updatedDate = $updatedDate;
+
+        return $this;
+    }
+
+
+}
