+++ symfony.lock
@@ -141,6 +141,21 @@
     "psr/log": {
         "version": "1.1.2"
     },
+    "ramsey/uuid": {
+        "version": "3.9.2"
+    },
+    "ramsey/uuid-doctrine": {
+        "version": "1.3",
+        "recipe": {
+            "repo": "github.com/symfony/recipes-contrib",
+            "branch": "master",
+            "version": "1.3",
+            "ref": "471aed0fbf5620b8d7f92b7a5ebbbf6c0945c27a"
+        },
+        "files": [
+            "config/packages/ramsey_uuid_doctrine.yaml"
+        ]
+    },
     "sensio/framework-extra-bundle": {
         "version": "5.2",
         "recipe": {
